+++ opensampl/vendors/microchip/tp4100.py
@@ -2,7 +2,7 @@
 
 import random
 from pathlib import Path
-from typing import ClassVar, Optional, Union
+from typing import ClassVar
 
 import click
 import pandas as pd
@@ -30,13 +30,13 @@ class MicrochipTP4100Probe(BaseProbe, RandomDataMixin):
         """Model for storing random data generation configurations as provided by CLI or YAML"""
 
         # Time series parameters
-        base_value: Optional[float] = Field(
+        base_value: float | None = Field(
             default_factory=lambda: random.uniform(-5e-7, 5e-7), description="random.uniform(-5e-7, 5e-7)"
         )
-        noise_amplitude: Optional[float] = Field(
+        noise_amplitude: float | None = Field(
             default_factory=lambda: random.uniform(1e-8, 5e-8), description="random.uniform(1e-8, 5e-8)"
         )
-        drift_rate: Optional[float] = Field(
+        drift_rate: float | None = Field(
             default_factory=lambda: random.uniform(-1e-10, 1e-10), description="random.uniform(-1e-10, 1e-10)"
         )
 
@@ -59,9 +59,9 @@ class MicrochipTP4100Probe(BaseProbe, RandomDataMixin):
         ]
         return base_options + vendor_options
 
-    def __init__(self, input_file: Union[str, Path]):
+    def __init__(self, input_file: str | Path, **kwargs: dict):
         """Initialize MicrochipTP4100 object given input_file and determines probe identity from file headers"""
-        super().__init__(input_file=input_file)
+        super().__init__(input_file=input_file, **kwargs)
         self.header = self.get_header()
         self.probe_key = ProbeKey(
             ip_address=self.header.get("host"), probe_id=self.header.get("probe_id", None) or "1-1"